PDFBOX-2576: make fields final; remove uneeded casts; remove public modifier from methods of non public class
git-svn-id: https://svn.apache.org/repos/asf/pdfbox/trunk@1670743 13f79535-47bb-0310-9956-ffa450edef68

+++ pdfbox/src/main/java/org/apache/pdfbox/pdmodel/graphics/shading/AxialShadingContext.java
@@ -132,6 +132,7 @@ public class AxialShadingContext extends ShadingContext implements PaintContext
      *
      * @return an array, index denotes the relative position, the corresponding
      * value is the color on the axial line
+     * @throws IOException if the color conversion fails.
      */
     private int[] calcColorTable() throws IOException
     {
@@ -145,7 +146,7 @@ public class AxialShadingContext extends ShadingContext implements PaintContext
         {
             for (int i = 0; i <= factor; i++)
             {
-                float t = domain[0] + d1d0 * i / (float) factor;
+                float t = domain[0] + d1d0 * i / factor;
                 float[] values = axialShadingType.evalFunction(t);
                 map[i] = convertToRGB(values);
             }
@@ -292,6 +293,8 @@ public class AxialShadingContext extends ShadingContext implements PaintContext
 
     /**
      * Returns the function.
+     *
+     * @throws java.io.IOException if we were not able to create the function.
      */
     public PDFunction getFunction() throws IOException
     {
